requirements.txt
amqp==2.6.1
arrow==0.17.0
asgiref==3.3.1
astroid==2.4.2
billiard==3.5.0.5
celery==4.1.1
cffi==1.14.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
Django==3.1.5
django-celery-results==2.0.0
djangorestframework==3.12.2
dnspython==1.16.0
eventlet==0.30.0
gevent==20.12.1
greenlet==0.4.17
huey==2.3.0
isort==5.7.0
Jinja2==2.11.2
kombu==4.6.11
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.3
prompt-toolkit==3.0.10
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.5
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
vine==1.3.0
wcwidth==0.2.5
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.2.0
